updated to correspond to book chapters

+++ ch07/AccountManagement/src/main/java/mjg/spring/dao/AccountDAO.java
@@ -17,16 +17,13 @@ package mjg.spring.dao;
 
 import mjg.spring.entities.Account;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountDAO {
-    int createAccount(double initialBalance);
-
+    int createAccount(BigDecimal initialBalance);
     Account findAccountById(int id);
-
     List<Account> findAllAccounts();
-
     void updateAccount(Account account);
-
     void deleteAccount(int id);
 }
