+++ ch07/AccountManagement/src/main/resources/schema.sql
@@ -1,2 +1,2 @@
 drop table accounts if exists;
-create table accounts (id integer primary key, balance double not null);
+create table accounts (id integer primary key, balance decimal not null);
